requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Data & ML
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
vectorbt==0.26.2
ta-lib==0.4.28
scipy==1.12.0

# Trading
alpaca-trade-api==3.0.2
alpaca-py==0.21.1
websocket-client==1.7.0

# Webull Official OpenAPI SDK
webull-python-sdk-core>=0.1.18
webull-python-sdk-trade>=0.1.18
webull-python-sdk-quotes-core>=0.1.18
webull-python-sdk-mdata>=0.1.18
webull-python-sdk-trade-events-core>=0.1.18
webull>=0.6.1

# Dashboard
streamlit==1.30.0
plotly==5.18.0

# LLM Intelligence
anthropic>=0.39.0
openai>=1.50.0

# Utilities
httpx==0.26.0
apscheduler==3.10.4
structlog==24.1.0
tenacity==8.2.3
joblib==1.3.2

# Auth & Security
bcrypt==4.1.2
cryptography==42.0.2

# DevOps
docker==7.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
